requirements.txt
Django==3.1.4
djangorestframework==3.12.2
django-model-utils==4.1.1
django-filter==2.4.0
Markdown==3.3.3
requests==2.25.1
pandas==1.2.0
